removing circular reference for SystemDict for py2.7, again, again
